third_party/WebKit/LayoutTests/editing/selection/select_all/select_all_in_contenteditable.html

<!doctype html>
<script src="../../../resources/testharness.js"></script>
<script src="../../../resources/testharnessreport.js"></script>
<script src="../../assert_selection.js"></script>
<div id="log"></div>
<script>
test(() => {
  assert_selection(
    'foo<div contenteditablele>|bar</div>baz',
    'selectAll',
    '^foo<div contenteditablele>bar</div>baz|');
}, '"Select All" when in an editable area selects the whole containing document');
</script>
